Imported from iceweasel_18.0.orig.tar.bz2

+++ build/mobile/remoteautomation.py
@@ -11,7 +11,7 @@ import shutil
 import subprocess
 
 from automation import Automation
-from devicemanager import NetworkTools
+from devicemanager import NetworkTools, DMError
 
 class RemoteAutomation(Automation):
     _devicemanager = None
@@ -80,13 +80,22 @@ class RemoteAutomation(Automation):
         return status
 
     def checkForCrashes(self, directory, symbolsPath):
-        dumpDir = tempfile.mkdtemp()
-        self._devicemanager.getDirectory(self._remoteProfile + '/minidumps/', dumpDir)
-        automationutils.checkForCrashes(dumpDir, symbolsPath, self.lastTestSeen)
-        try:
-          shutil.rmtree(dumpDir)
-        except:
-          print "WARNING: unable to remove directory: %s" % (dumpDir)
+        remoteCrashDir = self._remoteProfile + '/minidumps/'
+        if self._devicemanager.dirExists(remoteCrashDir):
+            dumpDir = tempfile.mkdtemp()
+            self._devicemanager.getDirectory(remoteCrashDir, dumpDir)
+            automationutils.checkForCrashes(dumpDir, symbolsPath,
+                                            self.lastTestSeen)
+            try:
+                shutil.rmtree(dumpDir)
+            except:
+                print "WARNING: unable to remove directory: %s" % dumpDir
+        else:
+            # As of this writing, the minidumps directory is automatically
+            # created when fennec (first) starts, so its lack of presence
+            # is a hint that something went wrong.
+            print "WARNING: No crash directory (%s) on remote " \
+                "device" % remoteCrashDir
 
     def buildCommandLine(self, app, debuggerInfo, profileDir, testURL, extraArgs):
         # If remote profile is specified, use that instead
@@ -164,27 +173,39 @@ class RemoteAutomation(Automation):
 
         @property
         def pid(self):
-            hexpid = self.dm.processExist(self.procName)
-            if (hexpid == None):
-                hexpid = "0x0"
-            return int(hexpid, 0)
+            pid = self.dm.processExist(self.procName)
+            # HACK: we should probably be more sophisticated about monitoring
+            # running processes for the remote case, but for now we'll assume
+            # that this method can be called when nothing exists and it is not
+            # an error
+            if pid is None:
+                return 0
+            return pid
 
         @property
         def stdout(self):
             """ Fetch the full remote log file using devicemanager and return just
                 the new log entries since the last call (as a multi-line string).
             """
-            t = self.dm.getFile(self.proc)
-            if t == None: return ''
-            newLogContent = t[self.stdoutlen:]
-            self.stdoutlen = len(t)
-            # Match the test filepath from the last TEST-START line found in the new
-            # log content. These lines are in the form:
-            # 1234 INFO TEST-START | /filepath/we/wish/to/capture.html\n
-            testStartFilenames = re.findall(r"TEST-START \| ([^\s]*)", newLogContent)
-            if testStartFilenames:
-                self.lastTestSeen = testStartFilenames[-1]
-            return newLogContent.strip('\n').strip()
+            if self.dm.fileExists(self.proc):
+                try:
+                    t = self.dm.pullFile(self.proc)
+                except DMError:
+                    # we currently don't retry properly in the pullFile
+                    # function in dmSUT, so an error here is not necessarily
+                    # the end of the world
+                    return ''
+                newLogContent = t[self.stdoutlen:]
+                self.stdoutlen = len(t)
+                # Match the test filepath from the last TEST-START line found in the new
+                # log content. These lines are in the form:
+                # 1234 INFO TEST-START | /filepath/we/wish/to/capture.html\n
+                testStartFilenames = re.findall(r"TEST-START \| ([^\s]*)", newLogContent)
+                if testStartFilenames:
+                    self.lastTestSeen = testStartFilenames[-1]
+                return newLogContent.strip('\n').strip()
+            else:
+                return ''
 
         @property
         def getLastTestSeen(self):
