+++ testing/mochitest/runtestsremote.py
@@ -200,10 +200,15 @@ class MochiRemote(Mochitest):
         self.remoteProfile = options.remoteTestRoot + "/profile"
         self._automation.setRemoteProfile(self.remoteProfile)
         self.remoteLog = options.remoteLogFile
+        self.localLog = options.logFile
 
     def cleanup(self, manifest, options):
-        self._dm.getFile(self.remoteLog, self.localLog)
-        self._dm.removeFile(self.remoteLog)
+        if self._dm.fileExists(self.remoteLog):
+            self._dm.getFile(self.remoteLog, self.localLog)
+            self._dm.removeFile(self.remoteLog)
+        else:
+            print "WARNING: Unable to retrieve log file (%s) from remote " \
+                "device" % self.remoteLog
         self._dm.removeDir(self.remoteProfile)
 
         if (options.pidFile != ""):
@@ -284,8 +289,11 @@ class MochiRemote(Mochitest):
         manifest = Mochitest.buildProfile(self, options)
         self.localProfile = options.profilePath
         self._dm.removeDir(self.remoteProfile)
-        if self._dm.pushDir(options.profilePath, self.remoteProfile) == None:
-            raise devicemanager.FileError("Unable to copy profile to device.")
+        try:
+            self._dm.pushDir(options.profilePath, self.remoteProfile)
+        except devicemanager.DMError:
+            print "Automation Error: Unable to copy profile to device."
+            raise
 
         options.profilePath = self.remoteProfile
         return manifest
@@ -296,8 +304,11 @@ class MochiRemote(Mochitest):
         options.profilePath = self.localProfile
         retVal = Mochitest.buildURLOptions(self, options, env)
         #we really need testConfig.js (for browser chrome)
-        if self._dm.pushDir(options.profilePath, self.remoteProfile) == None:
-            raise devicemanager.FileError("Unable to copy profile to device.")
+        try:
+            self._dm.pushDir(options.profilePath, self.remoteProfile)
+        except devicemanager.DMError:
+            print "Automation Error: Unable to copy profile to device."
+            raise
 
         options.profilePath = self.remoteProfile
         options.logFile = self.localLog
@@ -309,8 +320,12 @@ class MochiRemote(Mochitest):
           return "NO_CHROME_ON_DROID"
         path = '/'.join(parts[:-1])
         manifest = path + "/chrome/" + os.path.basename(filename)
-        if self._dm.pushFile(filename, manifest) == False:
-            raise devicemanager.FileError("Unable to install Chrome files on device.")
+        try:
+            self._dm.pushFile(filename, manifest)
+        except devicemanager.DMError:
+            print "Automation Error: Unable to install Chrome files on device."
+            raise
+
         return manifest
 
     def getLogFilePath(self, logFile):             
@@ -420,7 +435,10 @@ def main():
         mp.read(options.robocop)
         robocop_tests = mp.active_tests(exists=False)
 
-        fHandle = open("robotium.config", "w")
+        fHandle = tempfile.NamedTemporaryFile(suffix='.config',
+                                              prefix='robotium-',
+                                              dir=os.getcwd(),
+                                              delete=False)
         fHandle.write("profile=%s\n" % (mochitest.remoteProfile))
         fHandle.write("logfile=%s\n" % (options.remoteLogFile))
         fHandle.write("host=http://mochi.test:8888/tests\n")
@@ -428,18 +446,20 @@ def main():
         fHandle.close()
         deviceRoot = dm.getDeviceRoot()
       
-        # Note, we are pushing to /sdcard since we have this location hard coded in robocop
-        dm.removeFile("/sdcard/fennec_ids.txt")
-        dm.removeFile("/sdcard/robotium.config")
-        dm.pushFile("robotium.config", "/sdcard/robotium.config")
+        dm.removeFile(os.path.join(deviceRoot, "fennec_ids.txt"))
+        dm.removeFile(os.path.join(deviceRoot, "robotium.config"))
+        dm.pushFile(fHandle.name, os.path.join(deviceRoot, "robotium.config"))
+        os.unlink(fHandle.name)
         fennec_ids = os.path.abspath("fennec_ids.txt")
         if not os.path.exists(fennec_ids) and options.robocopIds:
             fennec_ids = options.robocopIds
-        dm.pushFile(fennec_ids, "/sdcard/fennec_ids.txt")
+        dm.pushFile(fennec_ids, os.path.join(deviceRoot, "fennec_ids.txt"))
         options.extraPrefs.append('robocop.logfile="%s/robocop.log"' % deviceRoot)
+        options.extraPrefs.append('browser.search.suggest.enabled=true')
+        options.extraPrefs.append('browser.search.suggest.prompted=true')
 
         if (options.dm_trans == 'adb' and options.robocopPath):
-          dm.checkCmd(["install", "-r", os.path.join(options.robocopPath, "robocop.apk")])
+          dm._checkCmd(["install", "-r", os.path.join(options.robocopPath, "robocop.apk")])
 
         appname = options.app
         retVal = None
@@ -449,9 +469,9 @@ def main():
                 continue
 
             options.app = "am"
-            options.browserArgs = ["instrument", "-w", "-e", "class"]
+            options.browserArgs = ["instrument", "-w", "-e", "deviceroot", deviceRoot, "-e", "class"]
             options.browserArgs.append("%s.tests.%s" % (appname, test['name']))
-            options.browserArgs.append("org.mozilla.roboexample.test/android.test.InstrumentationTestRunner")
+            options.browserArgs.append("org.mozilla.roboexample.test/%s.FennecInstrumentationTestRunner" % appname)
 
             try:
                 dm.recordLogcat()
